Move `PartialVersion` into submodule
